Always send join/part SMS
This was causing only one person to get the SMS messages, since we
record that the person left as soon as we hear about it, and we process
stanzas in order, so no one else would get notified.

If we turn out to need this kind of filter, we can do a per-person list
of who they think is in the room or something?