Room presence manager
We now store on a per-tel basis who they have seen join/part the room.
Joins/parts are stored based on what they have actually seen (via
debounce) and so should always map to SMS we actually sent.

If we get a join from someone we already thought was in the room, that
information is useless.  Do not send it.  Closes #34

We also update the global list in the same place, which I think is now
only used for the re-join on daemon restart.  This is related to #30,
but may not complete it yet.