Nuke many log lines
We were always logging effectively in debug mode, but with the amount of active
use we have this was both a possibly privacy issue and make uselessly big log
files.  I haven't looked at most of this in ages, so take it all out until
something calls for it.  Things mostly work these days.