Replace sendMessage, componentJid, from with sendText
It's funny, I built the "sendText" method for my cancellation logic as
an easy way for it to send some status messages to the user. After I
looked at it, though, I realized that's just what I always wanted. I'm
not sending any special messages back to the user, just text, and I'm
threading these arguments all the way through the logic just so I can
build a message from text at the bottom.

That's dumb, so now I'm just doing it at the top and injecting that.