Adhoc Session Cancellation and Timeout
This allows a user to bail-out without committing to anything. So they can type
"cancel" at any time to get out of the thing they're in the middle of and it'll
send the cancellation upstream to the provider, and then clear the session out
of our local set.

The same thing happens if the user disappeared part way through our
conversation, configured by a value in the code.
Maybe we could move this configuration out to the config file at some point,
but this is probably fine.