Adhoc Display List Field Values
The responder can optionally send us "current values" of fields, but
we're only displaying that information in the text field. This extends
that to the list types.

We also extended text fields to show descriptions. We wanted to extend
this to lists as well, but figuring out how to consolidate was a little
tricky. So I pulled some bits out, and we decided on a bit more of a
"question" vibe.

So now it looks like:

  Run Modes
  1. Single-User
  2. Non-Networked Multi-User
  3. Full Multi-User [Currently Selected]
  4. X-Window [Currently Selected]
  Which numbers?

and:

  Run State
  1. Active [Current Value]
  2. Inactive
  Which number?

and with a description it'd be:

  Run State
  (Which state should this run in?)
  1. Active [Current Value]
  2. Inactive
  Which number?

I felt it was clearer to have "Current Value" in the single case,
because having one element be selected isn't as clear when there's just
one, but that may just be me.

In the middle I also changed the text labels to look like:

  Enter First Name

  Enter Name
  (This is the name other people will see you by)

For fields labeled "First Name" and "Name" respectively.


Also, once I made the listOptionText method, and gave it a type
signature, I no longer need to specify the type of my decimal number in
the middle of my parser, the inference can figure it out now.