Adhoc JID Field
This is basically a text field, but it only accepts JIDs.
In a normal implementation this would pull up a convenient list of
contacts or something, but since we don't have any control like that,
this will have to do.

So we consume just the same as a text form, but with special validation
and instructions. That being said, I couldn't actually find a string
that wasn't accepted as a valid JID, so maybe it doesn't do anything
beyond just providing slightly different instruction.

Anyway, I adjusted untilParse to take (Text -> Maybe b) rather than
(Parser b) directly. Then I have the parse calls on the outside.
I think that's ok, in general.