Adhoc Hidden Field
This is super easy.
The only downside here is that it sends all the attributes of the field,
like the type and the label or whatever.

In practice, though, I assume no server impl will break because of that,
and it's so easy.

I could pull the value out and then repackage it, but if that's not
required then this is much simpler