Adhoc Boolean Field
This one is a bit trickier than some of the others, but still not too
bad. The main thing I have to do is turn the value the responder gives
us into something we can show to the user, and then interpret that
response as something we can give to the responder.

I use the same kinda weird parser for that where I just spell out all
the options

Seems to work ok...

It looks like:
  Is Enabled [Current value Yes]
  Yes or No?

and
  Is Not Enabled [Current value No]
  (Should it not not be disabled)
  Yes or No?

If there's a description.