Fix CI
jingle-xmpp isn't on hackage (yet?) so we need to install it from git.

Because we install everything from hackage in CI instead of using Debian
packages, installing jingle-xmpp first can cause a version of random not
compatible with some cheogram dependencies to get installed. Easiest fix for
that is to just install a known-good version first, then everything works.