Add Origin-ID to Messages from Adhoc Bot
We noticed that Dino was de-duping our messages, so if we typed "help"
two times in a row, the second appeared to hang.

Worst, the helper text on parse fails is also the same every time, so if
you typed something wrong, you'd get a message telling you how to fix
it. But if you typed the wrong thing again, even on a different question
in the same minute, you'd just get nothing back.

Not good!

So after looking in their code, it looks like it uses the origin-id from
XEP-0359 if present, so we're now adding one.

The facilitate that, I've switched the "sendMessage" method from STM ()
to (Unexceptional m) => m (), to allow me to generate UUIDs

On the one hand, m () is a bit carte-blanche, but on the other hand STM
was perhaps exposing too much of the underlying implementation.

It doesn't have to be writing to a queue, that's just what we've got.

I did also consider an alternative implementation where we have a TMVar
with a UUID in it, and spin up a new thread that just constantly tries
to fill the TMVar with a UUID, just so the STM could stay and we could
atomically pull the UUID and write to the chan, but in the end that was
likely overcomplicated for what I wanted to do.

So I did this instead.