Use Record for componentStanza context
componentStanza is a rediculous procedure with far too many branches, and as a
result it gets bits of context relavant to this branch or that one, but most
branches use only a few of the bits of context.  This change bundles the context
up into a record and only unpacks the fields used by a given branch.

While this does not fix the underlying design issues, it *is* more readable than
a long string of underscores that one must count through, and adding a new bit
of context will no longer require editing every single branch.  This seems like
a win.