Preserve file extesion through cacheOOB
Some users have MIME-blind tooling that really likes file extensions like the
olden days.  Doesn't hurt us to support that in some cases.  So, if the original
URL had an extension (something at the end ofter a . 4 chars or less) we append
it to the new URL.  Make sure your server is set up to rewrite these away!