Add Resource to Presence Queries
We had another case where Conversations expected a resource and hard-crashed
without one.

So when I hit the new "phone call" button we just enabled, it died.
Putting this in does technically change things for users, but likely not in a
way any of them would actually notice, and it makes Conversations die less.
So fine.