Support Backend Default Sip Proxies
Using XEP-0215 for a slightly unconventional purpose, we built code that
requests a sip service from the user's configured backend.

So that means the user can have its own sip server setup in Cheogram,
but if they don't then we can request it from the backend.
That means outbound calling will be set for anyone using a supported
backend, without us having to get all of them to run this command.

And if neither they nor their backend have sip configured, then there
just won't be a call button!