Respond with avatar contents when asked
If an avatar is specified in the config file, grab the needed metadata.  We
assume the avatar is image/png since the spec mostly requires at least that
codec to be present.  If we receive a pubsub request for exactly the node and
item that is this avatar, we return the base64 encoded content.