Allow backends to ask users to start registration
This can power, eg, a web flow that ends in asking Cheogram to finish up and
will power the new JMP flow.  The abuse potential is low because you don't get
free-form text, all that can happen is you can have Cheogram ask the user if
they want to register with your hostname.  If they user ignores it nothing
happens, the end.