Read from redis first, then TC -- write to both
If a key is in redis, use that value. Combine set and hash results from redis
and TC.  All writes go fully to both.

Redis keys are seperated by 0xFF instead of 0x0, the former is actually invalid
in UTF8 so is a fully safe seperator, and has the neat bonus of being used by
CBOR as a terminator (though the usage here does not make the keys valid CBOR).

Hashes are encoded so that we can store (Text, Maybe Text) -- a value of Just x
is stored UTF8 encoded, so if there are no Nothing values the hash looks like a
normal (Text, Text) would.  Nothing is stored as 0xF6, which is invalid UTF8
when alone so cannot be confused with valid text, and is the valid CBOR encoding
of null.

Enum are stored as the text representation of the corrosponding Int.  This is
only used on Bool in the codebase at the moment, so it will be 0 in Redis for
False, 1 for True.  This encoding is pretty common for integers in redis in
general.