Only store muc_membersonly if True
The default is False anyway if not present.

This handler is overbroad and catches any disco#info that isn't caught
elsewhere, but assumes this must be a result from a MUC.  That was probably true
when this was written, but it's certainly not safe and we can see in the DB that
there are muc_membersonly keys set to False for non-MUC JIDs.  This change is
safe and not invasive but will cut down on DB contents at least.