AdHoc Bot Continues on Result Forms
Previously it would display results and then be done, but I may want to
display resutlts and then continue, so I pulled the logic out of the
normal form-filling path and put it in result so it will continue
automatically if we're still executing.

This means that the bot will render back-to-back result forms as just a
dump of their output, but honestly that's probably fine.
What's the point in asking if they'd like to continue if I can just
continue until we hit a question we actually need the human to answer!