Interpret Invalid Open List Numbers as Strings
This is a bit risky maybe, but previously if I was in the admin menu and
there I wanted to search by a phone number (that didn't start with +) or
a customer ID, I couldn't.

This is because the parser would intepret "4873278943" as an option I
wanted out of the numbered list, but that wasn't an option, so it would
ask me to pick another number.

Now instead I say if it's not a valid list, just pass it through as
text.