Don't Forward Message Errors to the AdHoc Bot
These are in the messages that the user is sending the bot, not in the
commands its running on the other side.

In the worst case, this can lead to a feedback loop where we send a
message to a user, it errors back, so we send help text, it errors, we
send help text, etc, etc.

But more generally if something fails we should be safe to just ignore
it. Either someone will try something and we'll catch them later, or
it'll timeout later and they can start over.

There's no handling the error, because we don't really know what
happened. We just weren't able to communicate with the user, as a
service that does nothing but communicates with users...